Crate reth_nippy_jar

Immutable data store format.

Re-exports

pub use phf::PHFKey;

Modules

compression

filter

phf

Structs

DataReader

Manages the reading of snapshot data using memory-mapped files.

NippyJar

NippyJar is a specialized storage format designed for immutable data.

NippyJarCursor

Simple cursor implementation to retrieve data from NippyJar.

NippyJarWriter

Writer of NippyJar. Handles table data and offsets only.

Enums

NippyJarError

Errors associated with crate::NippyJar.

Traits

NippyJarHeader

A trait for the user-defined header of NippyJar.

Type Aliases

ColumnResult

Alias type for a column value wrapped in Result.