Struct reth_nippy_jar::compression::Decompressor

pub struct Decompressor<'a> { /* private fields */ }

Allows to decompress independently multiple blocks of data.

This reduces memory usage compared to calling decompress multiple times.

Implementations

impl Decompressor<'static>

pub fn new() -> Result<Decompressor<'static>, Error>

Creates a new zstd decompressor.

pub fn with_dictionary( dictionary: &[u8] ) -> Result<Decompressor<'static>, Error>

Creates a new zstd decompressor, using the given dictionary.

impl<'a> Decompressor<'a>

pub fn with_prepared_dictionary<'b>( dictionary: &'a DecoderDictionary<'b> ) -> Result<Decompressor<'a>, Error>
where 'b: 'a,

Creates a new decompressor using an existing DecoderDictionary.

Note that using a dictionary means that compression will need to use the same dictionary.

pub fn set_dictionary(&mut self, dictionary: &[u8]) -> Result<(), Error>

Changes the dictionary used by this decompressor.

Will affect future compression jobs.

Note that using a dictionary means that compression will need to use the same dictionary.

pub fn set_prepared_dictionary<'b>( &mut self, dictionary: &'a DecoderDictionary<'b> ) -> Result<(), Error>
where 'b: 'a,

Changes the dictionary used by this decompressor.

Note that using a dictionary means that compression will need to use the same dictionary.

pub fn decompress_to_buffer<C>( &mut self, source: &[u8], destination: &mut C ) -> Result<usize, Error>
where C: WriteBuf + ?Sized,

Deompress a single block of data to the given destination buffer.

Returns the number of bytes written, or an error if something happened (for instance if the destination buffer was too small).

pub fn decompress( &mut self, data: &[u8], capacity: usize ) -> Result<Vec<u8>, Error>

Decompress a block of data, and return the result in a Vec<u8>.

The decompressed data should be less than capacity bytes, or an error will be returned.

pub fn set_parameter(&mut self, parameter: DParameter) -> Result<(), Error>

Sets a decompression parameter for this decompressor.

pub fn window_log_max(&mut self, log_distance: u32) -> Result<(), Error>

Sets the maximum back-reference distance.

The actual maximum distance is going to be 2^log_distance.

This will need to at least match the value set when compressing.

pub fn include_magicbytes( &mut self, include_magicbytes: bool ) -> Result<(), Error>

Enables or disabled expecting the 4-byte magic header

Only available with the experimental feature.

This will need to match the settings used when compressing.

pub fn upper_bound(_data: &[u8]) -> Option<usize>

Get an upper bound on the decompressed size of data, if available

This can be used to pre-allocate enough capacity for decompress_to_buffer and is used by decompress to ensure that it does not over-allocate if you supply a large capacity.

Will return None if the upper bound cannot be determined or is larger than usize::MAX

Note that unless the experimental feature is enabled, this will always return None.