Trait reth_nippy_jar::phf::PerfectHashingFunction

pub trait PerfectHashingFunction: Serialize + for<'a> Deserialize<'a> {
    // Required methods
    fn set_keys<T: PHFKey>(&mut self, keys: &[T]) -> Result<(), NippyJarError>;
    fn get_index(&self, key: &[u8]) -> Result<Option<u64>, NippyJarError>;
}

Trait to build and query a perfect hashing function.

Required Methods

fn set_keys<T: PHFKey>(&mut self, keys: &[T]) -> Result<(), NippyJarError>

Adds the key set and builds the perfect hashing function.

fn get_index(&self, key: &[u8]) -> Result<Option<u64>, NippyJarError>

Get corresponding associated integer. There might be false positives.

Object Safety

This trait is not object safe.