Struct reth_nippy_jar::DataReader

pub struct DataReader { /* private fields */ }

Manages the reading of snapshot data using memory-mapped files.

Holds file and mmap descriptors of the data and offsets files of a snapshot.

Implementations

impl DataReader

pub fn new(path: impl AsRef<Path>) -> Result<Self, NippyJarError>

Reads the respective data and offsets file and returns DataReader.

pub fn offset(&self, index: usize) -> u64

Returns the offset for the requested data index

pub fn reverse_offset(&self, index: usize) -> Result<u64, NippyJarError>

Returns the offset for the requested data index starting from the end

pub fn offsets_count(&self) -> Result<usize, NippyJarError>

Returns total number of offsets in the file. The size of one offset is determined by the file itself.

pub fn offset_size(&self) -> u64

Returns number of bytes that represent one offset.

pub fn data(&self, range: Range<usize>) -> &[u8]

Returns the underlying data as a slice of bytes for the provided range.

pub fn size(&self) -> usize

Returns total size of data