Struct reth_nippy_jar::NippyJar

pub struct NippyJar<H = ()> { /* private fields */ }

NippyJar is a specialized storage format designed for immutable data.

Data is organized into a columnar format, enabling column-based compression. Data retrieval entails consulting an offset list and fetching the data from file via mmap.

PHF & Filters: For data membership verification, the filter field can be configured with algorithms like Bloom or Cuckoo filters. While these filters enable rapid membership checks, it’s important to note that they may yield false positives but not false negatives. Therefore, they serve as preliminary checks (eg. in by_hash queries) and should be followed by data verification on retrieval.

The phf (Perfect Hashing Function) and offsets_index fields facilitate the data retrieval process in for example by_hash queries. Specifically, the PHF converts a query, such as a block hash, into a unique integer. This integer is then used as an index in offsets_index, which maps to the actual data location in the offsets list. Similar to the filter, the PHF may also produce false positives but not false negatives, necessitating subsequent data verification.

Note: that the key (eg. BlockHash) passed to a filter and phf does not need to actually be stored.

Ultimately, the freeze function yields two files: a data file containing both the data and its configuration, and an index file that houses the offsets and offsets_index.

Implementations

impl NippyJar<()>

pub fn new_without_header(columns: usize, path: &Path) -> Self

Creates a new NippyJar without an user-defined header data.

pub fn load_without_header(path: &Path) -> Result<Self, NippyJarError>

Loads the file configuration and returns Self on a jar without user-defined header data.

pub fn uses_filters(&self) -> bool

Whether this NippyJar uses a InclusionFilters and Functions.

impl<H: NippyJarHeader> NippyJar<H>

pub fn new(columns: usize, path: &Path, user_header: H) -> Self

Creates a new NippyJar with a user-defined header data.

pub fn with_zstd(self, use_dict: bool, max_dict_size: usize) -> Self

Adds compression::Zstd compression.

pub fn with_lz4(self) -> Self

Adds compression::Lz4 compression.

pub fn with_cuckoo_filter(self, max_capacity: usize) -> Self

Adds filter::Cuckoo filter.

pub fn with_fmph(self) -> Self

Adds phf::Fmph perfect hashing function.

pub fn with_gofmph(self) -> Self

Adds phf::GoFmph perfect hashing function.

pub fn user_header(&self) -> &H

Gets a reference to the user header.

pub fn filter_size(&self) -> usize

Returns the size of inclusion filter

pub fn offsets_index_size(&self) -> usize

Returns the size of offsets index

pub fn compressor(&self) -> Option<&Compressors>

Gets a reference to the compressor.

pub fn compressor_mut(&mut self) -> Option<&mut Compressors>

Gets a mutable reference to the compressor.

pub fn load(path: &Path) -> Result<Self, NippyJarError>

Loads the file configuration and returns Self without deserializing filters related structures or the offset list.

The user must ensure the header type matches the one used during the jar’s creation.

pub fn load_filters(self) -> Result<Self, NippyJarError>

Loads filters into memory

pub fn data_path(&self) -> &Path

Returns the path for the data file

pub fn index_path(&self) -> PathBuf

Returns the path for the index file

pub fn offsets_path(&self) -> PathBuf

Returns the path for the offsets file

pub fn config_path(&self) -> PathBuf

Returns the path for the config file

pub fn open_data_reader(&self) -> Result<DataReader, NippyJarError>

Returns a DataReader of the data and offset file

pub fn prepare_compression( &mut self, columns: Vec<impl IntoIterator<Item = Vec<u8>>> ) -> Result<(), NippyJarError>

If required, prepares any compression algorithm to an early pass of the data.

pub fn prepare_index<T: PHFKey>( &mut self, values: impl IntoIterator<Item = ColumnResult<T>>, row_count: usize ) -> Result<(), NippyJarError>

Prepares beforehand the offsets index for querying rows based on values (eg. transaction hash). Expects values to be sorted in the same way as the data that is going to be later on inserted.

Currently collecting all items before acting on them.

pub fn freeze( &mut self, columns: Vec<impl IntoIterator<Item = ColumnResult<Vec<u8>>>>, total_rows: u64 ) -> Result<(), NippyJarError>

Writes all data and configuration to a file and the offset index to another.