Struct reth_nippy_jar::NippyJarCursor

pub struct NippyJarCursor<'a, H = ()> { /* private fields */ }

Simple cursor implementation to retrieve data from NippyJar.

Implementations

impl<'a, H: NippyJarHeader> NippyJarCursor<'a, H>

pub fn new(jar: &'a NippyJar<H>) -> Result<Self, NippyJarError>

pub fn with_reader( jar: &'a NippyJar<H>, reader: Arc<DataReader> ) -> Result<Self, NippyJarError>

pub fn jar(&self) -> &NippyJar<H>

Returns a reference to the related NippyJar

pub fn row_index(&self) -> u64

Returns current row index of the cursor

pub fn reset(&mut self)

Resets cursor to the beginning.

pub fn row_by_key( &mut self, key: &[u8] ) -> Result<Option<Vec<&'_ [u8]>>, NippyJarError>

Returns a row, searching it by a key used during NippyJar::prepare_index.

May return false positives.

Example usage would be querying a transactions file with a transaction hash which is NOT stored in file.

pub fn row_by_number( &mut self, row: usize ) -> Result<Option<Vec<&'_ [u8]>>, NippyJarError>

Returns a row by its number.

pub fn next_row(&mut self) -> Result<Option<Vec<&'_ [u8]>>, NippyJarError>

Returns the current value and advances the row.

pub fn row_by_key_with_cols( &mut self, key: &[u8], mask: usize ) -> Result<Option<Vec<&'_ [u8]>>, NippyJarError>

Returns a row, searching it by a key used during NippyJar::prepare_index by using a mask to only read certain columns from the row.

May return false positives.

Example usage would be querying a transactions file with a transaction hash which is NOT stored in file.

pub fn row_by_number_with_cols( &mut self, row: usize, mask: usize ) -> Result<Option<Vec<&'_ [u8]>>, NippyJarError>

Returns a row by its number by using a mask to only read certain columns from the row.

pub fn next_row_with_cols( &mut self, mask: usize ) -> Result<Option<Vec<&'_ [u8]>>, NippyJarError>

Returns the current value and advances the row.

Uses a mask to only read certain columns from the row.