Struct reth_nippy_jar::NippyJarWriter

pub struct NippyJarWriter<'a, H> { /* private fields */ }

Writer of NippyJar. Handles table data and offsets only.

Table data is written directly to disk, while offsets and configuration need to be flushed by calling commit().

Offset file layout

The first byte is the size of a single offset in bytes, m. Then, the file contains n entries, each with a size of m. Each entry represents an offset, except for the last entry, which represents both the total size of the data file, as well as the next offset to write new data to.

Data file layout

The data file is represented just as a sequence of bytes of data without any delimiters

Implementations

impl<'a, H: NippyJarHeader> NippyJarWriter<'a, H>

pub fn new(jar: &'a mut NippyJar<H>) -> Result<Self, NippyJarError>

pub fn append_rows( &mut self, column_values_per_row: Vec<impl IntoIterator<Item = ColumnResult<impl AsRef<[u8]>>>>, num_rows: u64 ) -> Result<(), NippyJarError>

Appends rows to data file. fn commit() should be called to flush offsets and config to disk.

column_values_per_row: A vector where each element is a column’s values in sequence, corresponding to each row. The vector’s length equals the number of columns.

pub fn append_column( &mut self, column: Option<ColumnResult<impl AsRef<[u8]>>> ) -> Result<(), NippyJarError>

Appends a column to data file. fn commit() should be called to flush offsets and config to disk.

pub fn prune_rows(&mut self, num_rows: usize) -> Result<(), NippyJarError>

Prunes rows from data and offsets file and updates its configuration on disk

pub fn commit(&mut self) -> Result<(), NippyJarError>

Commits configuration and offsets to disk. It drains the internal offset list.